Removed call to IsNullOrWhiteSpace on pszDelimiters in Split function
Since strings can also be whitespaces separated (tab-delimited, for
example, or LF-delimited or space delimited) the idea is that if we
insist that the delimiter string not contain whitespace, then that
defeats the point.  So, instead we check for whether the string pointer
is NULL or the first element points to '\0'.